refactor: move grid frontend files into component subfolder
