+++ vaadin-grid-flow-parent/vaadin-grid-flow/src/main/resources/META-INF/frontend/vaadin-grid/treeGridConnector.ts
@@ -1,5 +1,5 @@
 import './gridConnector.ts';
-import type { FlowTreeGrid } from './vaadin-grid/vaadin-grid-types.js';
+import type { FlowTreeGrid } from './vaadin-grid-types.js';
 
 /**
  * treeGridConnector is a communication layer between TreeGrid's flow component
